+++ apps/modernization-ui/src/pages/advancedSearch/components/patientSearch/ContactForm.tsx
@@ -1,7 +1,7 @@
 import { Grid } from '@trussworks/react-uswds';
 import { Input } from '../../../../components/FormInputs/Input';
 import { Controller } from 'react-hook-form';
-import { validate as validatePhoneNumber } from 'validation/phone/search';
+import { validate as validatePhoneNumber } from 'validation/search';
 
 export const ContactForm = ({ control, errors }: any) => {
     return (
