+++ apps/modernization-ui/src/pages/advancedSearch/components/patientSearch/PatientSearch.tsx
@@ -10,7 +10,7 @@ import { AddressForm } from './AddressForm';
 import { ContactForm } from './ContactForm';
 import { EthnicityForm } from './EthnicityForm';
 import { IDForm } from './IdForm';
-import { validate as validatePhoneNumber } from 'validation/phone/search';
+import { validate as validatePhoneNumber } from 'validation/search';
 import { validateZipCode } from '../../../../utils/ZipValidation';
 
 type PatientSearchProps = {
