+++ apps/modernization-ui/src/validation/search/SearchValidator.spec.ts
