+++ apps/modernization-ui/src/validation/search/SearchValidator.ts
